% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_run_ndr}
\alias{collect_run_ndr}
\title{Create native python script and move input files to a clean directory}
\usage{
collect_run_ndr(args, conda_path, conda_env, out_dir = "workspace_temp")
}
\arguments{
\item{args}{named list of ndr arguments}

\item{conda_path}{path to conda executable}

\item{conda_env}{path to conda environment}

\item{out_dir}{path to temporary working directory}
}
\description{
Create native python script and move input files to a clean directory
}
\examples{
\dontrun{
collect_run_ndr(ndr_testdata_args,
conda_path = paste0("/home/", Sys.info()[["user"]], "/anaconda3/bin"),
conda_env = paste0("/home/", Sys.info()[["user"]], "/Documents/Science/Models/invest/env")
)
}
}
