% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{invest}
\alias{invest}
\title{InVEST object}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
invest
}
\description{
Uses the reticulate framework to access the InVEST API.
}
\details{
The InVEST Python library is exposed through the `invest` object.

In this package, use the `$` operator wherever you see the `.` operator
used in Python.
}
\keyword{datasets}
