% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr.R
\name{ndr}
\alias{ndr}
\title{Nutrient Delivery Ratio (NDR) Model}
\usage{
ndr(args, overwrite = FALSE)
}
\arguments{
\item{args}{named list of model arguments}

\item{overwrite}{boolean}
}
\description{
Nutrient Delivery Ratio (NDR) Model
}
\examples{
\dontrun{

ndr(ndr_testdata_args)

}
}
