% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr.R
\name{ndr_p_export_total}
\alias{ndr_p_export_total}
\title{Calculate total P export}
\usage{
ndr_p_export_total(output_folder)
}
\arguments{
\item{output_folder}{path to ndr output folder

# @importFrom raster raster cellStats}
}
\description{
Calculate total P export
}
\examples{
\dontrun{
ndr_p_export_total("workspace")
}
}
