% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr.R
\name{ndr_testdata_args}
\alias{ndr_testdata_args}
\title{Test NDR arguments using included data}
\usage{
ndr_testdata_args()
}
\description{
Test NDR arguments using included data
}
\examples{
\dontrun{
ndr_testdata_args()
}
}
