% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr.R
\name{raster_extent_match}
\alias{raster_extent_match}
\title{Preflight check of raster extent matching}
\usage{
raster_extent_match(lulc_path, dem_path, runoff_proxy_path)
}
\description{
Preflight check of raster extent matching
}
\examples{
\dontrun{
raster_extent_match(ndr_testdata_args()$lulc_path, ndr_testdata_args()$dem_path, ndr_testdata_args()$runoff_proxy_path)
}
}
