% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr.R
\name{summarize_inputs_ndr}
\alias{summarize_inputs_ndr}
\title{Summarize NDR inputs}
\usage{
summarize_inputs_ndr(folder_path, args = NULL)
}
\arguments{
\item{folder_path}{path to a folder containing ndr input files}

\item{args}{optional. named list of ndr args}
}
\value{
a data.frame where each row is a lulc type and there are columns
for:
the number of cells of each type, the percent of each type, and the product
of the number of each type and load_p from the biophys table
}
\description{
Summarize NDR inputs
}
\details{
Assumes that the lulc raster has "lulc" in the file name.
}
\examples{
\dontrun{
summarize_inputs_ndr(folder_path = "inst/extdata/NDR", args = ndr_testdata_args())
}
}
