% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr.R
\name{summarize_outputs_ndr}
\alias{summarize_outputs_ndr}
\title{Summarize NDR outputs}
\usage{
summarize_outputs_ndr(folder_path)
}
\arguments{
\item{folder_path}{path to a folder containing ndr output files}
}
\description{
Summarize NDR outputs
}
\examples{
\dontrun{
output_summary <- summarize_outputs_ndr(folder_path = "workspace")
}
}
