% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_run_ndr}
\alias{collect_run_ndr}
\title{Create native python script and move input files to a clean directory}
\usage{
collect_run_ndr(
  args,
  out_dir = "workspace_temp",
  symlink = FALSE,
  conda_path = NULL,
  conda_env = NULL
)
}
\arguments{
\item{args}{named list of ndr arguments}

\item{out_dir}{path to temporary working directory}

\item{symlink}{symlink input files? FALSE copies files instead}

\item{conda_path}{optional. path to conda executable}

\item{conda_env}{optional. path to conda environment}
}
\description{
Create native python script and move input files to a clean directory
}
\examples{
\dontrun{
unlink("workspace_temp", recursive = TRUE)
collect_run_ndr(ndr_testdata_args())
collect_run_ndr(ndr_testdata_args(), symlink = TRUE)
}
}
