% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_invest}
\alias{install_invest}
\title{Install InVEST and its dependencies}
\usage{
install_invest(restart_session = TRUE)
}
\arguments{
\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}
}
\description{
Install InVEST and its dependencies
}
\examples{
\dontrun{
rinvest::install_invest()
}
}
