% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ndr}
\alias{ndr}
\title{Nutrient Delivery Ratio (NDR) Model}
\usage{
ndr(args, overwrite = FALSE)
}
\arguments{
\item{args}{named list of model arguments}

\item{overwrite}{boolean}
}
\description{
Nutrient Delivery Ratio (NDR) Model
}
\examples{
\dontrun{
data_dir <- system.file("extdata/NDR", package = "rinvest")
args <- list(
  "workspace_dir" = "workspace",
  "dem_path" = paste0(data_dir, "/DEM_gura.tif"),
  "lulc_path" = paste0(data_dir, "/land_use_gura.tif"),
  "runoff_proxy_path" = paste0(data_dir, "/precipitation_gura.tif"),
  "watersheds_path" = paste0(data_dir,  "/watershed_gura.shp"),
  "biophysical_table_path" = paste0(data_dir,  "/biophysical_table_gura.csv"),
  "calc_p" = TRUE,
  "calc_n" = FALSE,
  "threshold_flow_accumulation" = 1000,
  "k_param" = 2,
  "subsurface_eff_p" = 0.5,
  "subsurface_critical_length_p" = 25
)

ndr(args)

}
}
