% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr.R
\name{ndr_p_export_total}
\alias{ndr_p_export_total}
\title{Calculate total P export}
\usage{
ndr_p_export_total(ndr_output)
}
\arguments{
\item{ndr_output}{output of the ndr function. vector of file paths.}
}
\description{
Calculate total P export
}
