% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndr.R
\name{ndr_testdata_args}
\alias{ndr_testdata_args}
\title{Test NDR arguments using included data}
\usage{
ndr_testdata_args()
}
\description{
Test NDR arguments using included data
}
\examples{
\dontrun{
ndr_testdata_args()

# $workspace_dir
# [1] "workspace"

# $dem_path
# [1] "~/rinvest/inst/extdata/NDR/DEM_gura.tif"

# $lulc_path
# [1] "~/rinvest/inst/extdata/NDR/land_use_gura.tif"

# $runoff_proxy_path
# [1] "~/rinvest/inst/extdata/NDR/precipitation_gura.tif"

# $watersheds_path
# [1] "~/rinvest/inst/extdata/NDR/watershed_gura.shp"

# $biophysical_table_path
# [1] "~/rinvest/inst/extdata/NDR/biophysical_table_gura.csv"

# $calc_p
# [1] TRUE

# $calc_n
# [1] FALSE

# $threshold_flow_accumulation
# [1] 1000

# $k_param
# [1] 2

# $subsurface_eff_p
# [1] 0.5

# $subsurface_critical_length_p
# [1] 25

# $subsurface_eff_n
# [1] 0

# $subsurface_critical_length_n
# [1] 0
}
}
