% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinvest-package.R
\docType{package}
\name{rinvest-package}
\alias{rinvest}
\alias{rinvest-package}
\title{rinvest: InVEST: ecosystem service and valuation models}
\description{
Wrapper for the 'natcap.invest` python package
  <https://github.com/natcap/invest>, with example data for running 
  the Nutrient Delivery Ratio (NDR) model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsta/rinvest}
  \item Report bugs at \url{https://github.com/jsta/rinvest/issues}
}

}
\author{
\strong{Maintainer}: Joseph Stachelek \email{stachel2@msu.edu} (\href{https://orcid.org/0000-0002-5924-2464}{ORCID})

}
\keyword{internal}
